Consolidate utilities tests under unittests directory and fix test issues
- Remove utilities/tests directory and consolidate all tests under utilities/unittests
- Fix test_must_gather.py mock issues by correcting mock paths and removing global module mocking
- Update test_os_utils.py with correct function names and signatures
- Add test_ssp.py with proper mocking and function signatures
- Fix test_logger.py class naming to avoid pytest warnings
- Update coverage configuration to exclude conftest.py from coverage reports
- All 256 utilities tests now pass
