--- utilities/tests/TODO.md
@@ -1,15 +1,0 @@
-# CNV Utilities Testing TODO
-
-## 🔄 Remaining Work
-
-- data_collector.py
-- monitoring.py
-- os_utils.py
-- ssp.py
-- hco.py
-- infra.py
-- network.py
-- operator.py
-- pytest_utils.py
-- storage.py
-- virt.py
