--- utilities/tests/pytest.ini
@@ -1,31 +1,0 @@
-[pytest]
-# Independent pytest configuration for utilities tests
-addopts =
-    -v
-    --tb=short
-    --strict-markers
-    --disable-warnings
-    -p no:warnings
-    --no-header
-
-testpaths = .
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-
-# Markers specific to utilities tests
-markers =
-    unit: Unit tests that don't require external resources
-    integration: Tests that may require mocked external resources
-    slow: Tests that take longer to execute
-    mock_k8s: Tests that mock Kubernetes/OpenShift resources
-    optionalhook: pytest-html optional hook markers
-
-# Logging
-log_cli = false
-log_cli_level = INFO
-log_cli_format = %(asctime)s [%(levelname)s] %(message)s
-log_cli_date_format = %Y-%m-%d %H:%M:%S
-
-# Timeout for tests
-timeout = 300
