+++ utilities/unittests/TODO.md
@@ -2,14 +2,9 @@
 
 ## 🔄 Remaining Work
 
-- data_collector.py
-- monitoring.py
-- os_utils.py
-- ssp.py
 - hco.py
 - infra.py
 - network.py
 - operator.py
-- pytest_utils.py
 - storage.py
 - virt.py
