+++ utilities/unittests/conftest.py
@@ -6,6 +6,9 @@ from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
+from ocp_resources import resource
+
+import utilities
 
 os.environ["OPENSHIFT_VIRTUALIZATION_TEST_IMAGES_ARCH"] = "x86_64"
 
@@ -13,14 +16,26 @@ os.environ["OPENSHIFT_VIRTUALIZATION_TEST_IMAGES_ARCH"] = "x86_64"
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
 # Mock get_client to prevent K8s API calls
-from ocp_resources import resource  # noqa: E402
 
 resource.get_client = lambda: MagicMock()
 
 # Create mock modules to break circular imports
-sys.modules["utilities.data_collector"] = MagicMock()
-sys.modules["utilities.data_collector"].get_data_collector_base_directory = lambda: "/tmp/data"  # type: ignore
-sys.modules["utilities.data_collector"].collect_alerts_data = MagicMock()  # type: ignore
+# Set up mock modules before any imports
+mock_hco = MagicMock()
+mock_infra = MagicMock()
+mock_data_collector = MagicMock()
+mock_data_collector.get_data_collector_base_directory = MagicMock(return_value="/tmp/data")
+mock_data_collector.get_data_collector_base = MagicMock(return_value="/tmp/data/")
+
+sys.modules["utilities.hco"] = mock_hco
+sys.modules["utilities.infra"] = mock_infra
+sys.modules["utilities.data_collector"] = mock_data_collector
+
+# Also set them as attributes of the utilities module for tests that need them
+
+utilities.hco = mock_hco  # type: ignore[attr-defined]
+utilities.infra = mock_infra  # type: ignore[attr-defined]
+utilities.data_collector = mock_data_collector  # type: ignore[attr-defined]
 
 
 # Mock fixtures for common dependencies
@@ -30,6 +45,20 @@ def mock_get_client(monkeypatch):
     mock_client = MagicMock()
     monkeypatch.setattr("ocp_resources.resource.get_client", lambda: mock_client)
     return mock_client
+
+
+@pytest.fixture(autouse=True)
+def setup_py_config():
+    """Setup py_config for tests that need data_collector configuration"""
+    from pytest_testconfig import config as py_config
+
+    # Ensure data_collector config is set up
+    if "data_collector" not in py_config:
+        py_config["data_collector"] = {"data_collector_base_directory": "/tmp/data"}
+
+    yield
+
+    # Cleanup - restore original state if needed
 
 
 @pytest.fixture(autouse=True)
@@ -118,3 +147,121 @@ def mock_logger():
 
     # Restore original getLogger
     logging.getLogger = original_get_logger
+
+
+# Shared utility functions for data_collector tests
+def get_data_collector_base_logic(base_dir=None):
+    """Shared function simulating get_data_collector_base logic"""
+    if base_dir:
+        base_path = base_dir
+    elif os.environ.get("CNV_TESTS_CONTAINER"):
+        base_path = "/data"
+    else:
+        base_path = os.getcwd()
+
+    base_path = os.path.normpath(os.path.expanduser(base_path))
+    if not base_path.endswith(os.sep):
+        base_path = f"{base_path}{os.sep}"
+
+    return base_path
+
+
+@pytest.fixture
+def mock_write_to_file_logic():
+    """Fixture providing write_to_file logic function"""
+
+    def write_to_file_logic(file_name, content, base_directory, mode="w"):
+        """Simulate write_to_file logic"""
+        os.makedirs(base_directory, exist_ok=True)
+        file_path = os.path.join(base_directory, file_name)
+
+        try:
+            with open(file_path, mode) as fd:
+                fd.write(content)
+            return True
+        except Exception:
+            return False
+
+    return write_to_file_logic
+
+
+@pytest.fixture
+def mock_set_data_collector_values_logic():
+    """Fixture providing set_data_collector_values logic function"""
+
+    def set_data_collector_values_logic(base_dir=None, get_base_func=None):
+        """Simulate set_data_collector_values logic"""
+        base_path = get_base_func(base_dir=base_dir) if get_base_func else "/default/base/"
+        return {"data_collector_base_directory": f"{base_path}tests-collected-info"}
+
+    return set_data_collector_values_logic
+
+
+@pytest.fixture
+def mock_get_data_collector_dir_logic():
+    """Fixture providing get_data_collector_dir logic function"""
+
+    def get_data_collector_dir_logic(config_dict):
+        """Simulate get_data_collector_dir logic"""
+        return config_dict.get(
+            "collector_directory",
+            config_dict["data_collector_base_directory"],
+        )
+
+    return get_data_collector_dir_logic
+
+
+@pytest.fixture
+def mock_os_images():
+    """Fixture providing mock OS image classes for testing"""
+    # Mock RHEL class
+    mock_rhel_class = MagicMock()
+    mock_rhel_class.LATEST_RELEASE_STR = "rhel-9.6.qcow2"
+    mock_rhel_class.DEFAULT_DV_SIZE = "20Gi"
+    mock_rhel_class.RHEL7_9_IMG = "rhel-7.9.qcow2"
+    mock_rhel_class.RHEL8_10_IMG = "rhel-8.10.qcow2"
+    mock_rhel_class.RHEL9_5_IMG = "rhel-9.5.qcow2"
+    mock_rhel_class.RHEL9_6_IMG = "rhel-9.6.qcow2"
+    mock_rhel_class.DIR = "cnv-tests/rhel-images"
+
+    # Mock Windows class
+    mock_windows_class = MagicMock()
+    mock_windows_class.LATEST_RELEASE_STR = "win2k25.qcow2"
+    mock_windows_class.DEFAULT_DV_SIZE = "60Gi"
+    mock_windows_class.WIN10_IMG = "win10.qcow2"
+    mock_windows_class.WIN11_IMG = "win11.qcow2"
+    mock_windows_class.WIN2k16_IMG = "win2k16.qcow2"
+    mock_windows_class.WIN2k19_IMG = "win2k19.qcow2"
+    mock_windows_class.WIN2022_IMG = "win2022.qcow2"
+    mock_windows_class.WIN2k25_IMG = "win2k25.qcow2"
+    mock_windows_class.DIR = "cnv-tests/windows-images"
+    mock_windows_class.UEFI_WIN_DIR = "cnv-tests/windows-uefi-images"
+
+    # Mock Fedora class
+    mock_fedora_class = MagicMock()
+    mock_fedora_class.LATEST_RELEASE_STR = "fedora-41.qcow2"
+    mock_fedora_class.DEFAULT_DV_SIZE = "20Gi"
+    mock_fedora_class.FEDORA41_IMG = "fedora-41.qcow2"
+    mock_fedora_class.DIR = "cnv-tests/fedora-images"
+
+    # Mock CentOS class
+    mock_centos_class = MagicMock()
+    mock_centos_class.LATEST_RELEASE_STR = "centos-stream-9.qcow2"
+    mock_centos_class.DEFAULT_DV_SIZE = "20Gi"
+    mock_centos_class.CENTOS_STREAM_9_IMG = "centos-stream-9.qcow2"
+    mock_centos_class.DIR = "cnv-tests/centos-images"
+
+    # Mock Images container
+    mock_images = MagicMock()
+    mock_images.Rhel = mock_rhel_class
+    mock_images.Windows = mock_windows_class
+    mock_images.Fedora = mock_fedora_class
+    mock_images.Centos = mock_centos_class
+
+    return {
+        "images": mock_images,
+        "rhel": mock_rhel_class,
+        "windows": mock_windows_class,
+        "fedora": mock_fedora_class,
+        "centos": mock_centos_class,
+    }
